% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_issues.R
\name{fetch_issue_page}
\alias{fetch_issue_page}
\title{Fetch Issues from Jira Cloud - using URL}
\usage{
fetch_issue_page(query, username, token)
}
\arguments{
\item{query}{JQL Query}

\item{username}{Jira cloud username}

\item{token}{Jira cloud API token}
}
\value{
A dataframe of the query results
}
\description{
This function takes a JQL query and authentican artifacts and
returns a dataframe of the issues with a small number of key fields.
}
