% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_issues.R
\name{fix_empty}
\alias{fix_empty}
\title{Helper function to blank out empyt results}
\usage{
fix_empty(x)
}
\arguments{
\item{x}{thing to check if it's empty}
}
\value{
x or null
}
\description{
Helper function to blank out empyt results
}
