% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_issues.R
\name{get_issues}
\alias{get_issues}
\title{Fetch the raw results}
\usage{
get_issues(query, username, token)
}
\arguments{
\item{query}{JQL results}

\item{username}{authentication username}

\item{token}{API token}
}
\value{
A empty dataframe
}
\description{
Fetch the raw results
}
