% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_issues.R
\name{jqlQuery}
\alias{jqlQuery}
\title{Fetch Issues from Jira Cloud}
\usage{
jqlQuery(site, jql, username, token)
}
\arguments{
\item{site}{The Jira Cloud site to run the query against}

\item{jql}{JQL Query}

\item{username}{Jira cloud username}

\item{token}{Jira cloud API token}
}
\value{
A dataframe of the query results
}
\description{
This function takes a JQL query and authentican artifacts and
returns a dataframe of the issues with a small number of key fields.
}
