% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_issues.R
\name{populate_dataframe}
\alias{populate_dataframe}
\title{Populate data frame with results from JQL query}
\usage{
populate_dataframe(JSON_results, destination_dataframe)
}
\arguments{
\item{JSON_results}{JQL results}

\item{destination_dataframe}{The dataframe to write the results to}
}
\value{
A dataframe of the query results
}
\description{
Populate data frame with results from JQL query
}
