% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_issues.R
\name{prep_blank_dataframe}
\alias{prep_blank_dataframe}
\title{Make a blank dataframe that is the right size to put the JQL results into}
\usage{
prep_blank_dataframe(results)
}
\arguments{
\item{results}{JQL results}
}
\value{
A empty dataframe
}
\description{
Make a blank dataframe that is the right size to put the JQL results into
}
